% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_load_data.R
\name{fcs_load_data}
\alias{fcs_load_data}
\title{Load Example FCS Files Bundled with FCSimple}
\usage{
fcs_load_data()
}
\value{
A character vector of full file paths to each .fcs file found in
\code{system.file("data", package = "FCSimple")}.
}
\description{
Retrieves the file paths of all .fcs files included in the FCSimple
package’s internal \verb{data/} directory. Useful for quickly accessing
example or test files distributed with the package.
}
\examples{
\dontrun{
# List all example FCS files shipped with FCSimple
files <- FCSimple::fcs_load_data()

# Use them to join and preprocess
joined <- FCSimple::fcs_join(files)
}

}
\seealso{
FCSimple::fcs_join, utils::system.file
}
