% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_gating.R
\name{fcs_plot_cells}
\alias{fcs_plot_cells}
\title{Plot Gated Cells on SSC vs FSC}
\usage{
fcs_plot_cells(object, gate_tree, gate_name = "cells", downsample_size = 1e+05)
}
\arguments{
\item{object}{An “fcs_gating_object” containing \code{gate_trees[[gate_tree]][[gate_name]]}.}

\item{gate_tree}{Character; name of the gating tree (default “tree1”).}

\item{gate_name}{Character; name of the gate branch to plot (default “cells”).}

\item{downsample_size}{Integer; maximum points to plot (default 100000).}
}
\value{
A ggplot2 object.
}
\description{
Scatter‐plots all events in gray and gated events in blue, overlaying
the convex hull in red.
}
\examples{
\dontrun{
  go <- fcs_gate_cells(gating_obj, gate_name = "cells")
  p <- fcs_plot_cells(go, gate_name = "cells")
  print(p)
}

}
\seealso{
fcs_gate_cells, ggplot2::ggplot
}
